[REFACTOR] provider package 수정
