+++ src/main/java/store/ckin/auth/provider/JwtProvider.java
@@ -1,4 +1,4 @@
-package store.ckin.auth.util;
+package store.ckin.auth.provider;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
