+++ src/main/java/store/ckin/auth/provider/MemberAuthenticationProvider.java
@@ -1,4 +1,4 @@
-package store.ckin.auth.member.provider;
+package store.ckin.auth.provider;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationProvider;
