+++ src/main/java/store/ckin/auth/token/service/impl/TokenServiceImpl.java
@@ -7,7 +7,7 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import store.ckin.auth.token.service.TokenService;
-import store.ckin.auth.util.JwtProvider;
+import store.ckin.auth.provider.JwtProvider;
 
 /**
  * TokenService 의 구현체 입니다.
